requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.12.0
ucimlrepo>=0.0.3
xgboost>=1.7.0
streamlit-reveal-slides>=0.1.0
